+++ weblate/trans/models/component.py
@@ -26,6 +26,7 @@ from collections import Counter
 from contextlib import contextmanager
 from copy import copy
 from glob import glob
+from itertools import chain
 from typing import List, Optional
 from urllib.parse import urlparse
 
@@ -630,16 +631,33 @@ class Component(FastDeleteMixin, models.Model, URLMixin, PathMixin):
         """Installs automatically enabled addons from file format."""
         from weblate.addons.models import ADDONS
 
-        for name, configuration in self.file_format_cls.autoaddon.items():
-            try:
-                addon = ADDONS[name]
-            except KeyError:
-                self.log_warning("could not enable addon %s", name)
+        for name, configuration in chain(
+            self.file_format_cls.autoaddon.items(), settings.DEFAULT_ADDONS.items()
+        ):
+            if self.addon_set.filter(name=name).exists():
+                self.log_warning("could not enable addon %s, already installed", name)
                 continue
-            self.log_info("enabling addon %s", name)
 
-            if addon.can_install(self, None):
-                addon.create(self, configuration=configuration)
+            try:
+                addon = ADDONS[name]()
+            except KeyError:
+                self.log_warning("could not enable addon %s, not found", name)
+                continue
+
+            if addon.has_settings:
+                form = addon.get_add_form(self, data=configuration)
+                if not form.is_valid():
+                    self.log_warning(
+                        "could not enable addon %s, invalid settings", name
+                    )
+                    continue
+
+            if not addon.can_install(self, None):
+                self.log_warning("could not enable addon %s, not compatible", name)
+                continue
+
+            self.log_info("enabling addon %s", name)
+            addon.create(self, configuration=configuration)
 
     @contextmanager
     def lock(self):